slighlty simpler api calls

+++ include/constants.hpp
@@ -21,7 +21,7 @@ const float UPDATE_DELTA = 1.0f / 60.0f;
 
 // --- Physics Collision Categories ---
 // For simplicity in this setup, we'll use const.
-const uint16_t CATEGORY_PLAYER = 0x0001;
-const uint16_t CATEGORY_WORLD  = 0x0002;
+const uint64_t CATEGORY_PLAYER = 0x0001;
+const uint64_t CATEGORY_WORLD  = 0x0002;
 
 #endif // CONSTANTS_HPP
